refactor(daytona): split types.py into types.py + result_types.py
Move recursive-task, result, and rollout-summary types (Sections 4-5)
into a new result_types.py (~563 LOC). Keep serialization helpers,
budget/sandbox-spec, and prompt/context types in types.py (~441 LOC).

All moved symbols are re-exported from types.py so existing imports
remain backwards-compatible.
